refactor: 歌词播放器 base 部分集中到统一目录
