+++ packages/core/src/lyric-player/base/player.ts
@@ -5,15 +5,15 @@ import type {
 	LyricLine,
 	LyricWord,
 	OptimizeLyricOptions,
-} from "../interfaces.ts";
-import styles from "../styles/lyric-player.module.css";
-import { eqSet } from "../utils/eq-set.ts";
-import { isCJK } from "../utils/is-cjk.ts";
-import { optimizeLyricLines } from "../utils/optimize-lyric.ts";
-import { Spring, type SpringParams } from "../utils/spring.ts";
+} from "#src/interfaces.ts";
+import styles from "#styles/lyric-player.module.css";
+import { eqSet } from "#utils/eq-set.ts";
+import { optimizeLyricLines } from "#utils/optimize-lyric.ts";
+import type { SpringParams } from "#utils/spring.ts";
 import { BottomLineEl } from "./bottom-line.ts";
-import { InterludeDots } from "./dom/interlude-dots.ts";
-import { LyricLineRenderMode, MaskObsceneWordsMode } from "./index.ts";
+import { InterludeDots } from "../dom/interlude-dots.ts";
+import { LyricLineRenderMode, MaskObsceneWordsMode } from "./fixures.ts";
+import type { LyricLineBase } from "./line.ts";
 
 /**
  * 歌词播放器的基类，已经包含了有关歌词操作和排版的功能，子类需要为其实现对应的显示展示操作
@@ -1188,92 +1188,4 @@ export abstract class LyricPlayerBase
 		window.removeEventListener("pageshow", this.onPageShow);
 		window.removeEventListener("pagehide", this.onPageHide);
 	}
-}
-
-interface LineTransforms {
-	posY: Spring;
-	scale: Spring;
-}
-
-/**
- * 所有标准歌词行的基类
- * @internal
- */
-export abstract class LyricLineBase extends EventTarget implements Disposable {
-	protected top = 0;
-	protected scale = 1;
-	protected blur = 0;
-	protected opacity = 1;
-	protected delay = 0;
-	readonly lineTransforms: LineTransforms = {
-		posY: new Spring(0),
-		scale: new Spring(100),
-	};
-
-	/**
-	 * 用于 CJK 词语边界检测的分词器
-	 */
-	static readonly wordSegmenter: Intl.Segmenter | null =
-		typeof Intl !== "undefined" && Intl.Segmenter
-			? new Intl.Segmenter(undefined, {
-					granularity: "word",
-				})
-			: null;
-
-	/**
-	 * Unicode 标准的全局 Grapheme Cluster 分词器
-	 * 用于正确处理 emoji、复合字符等
-	 */
-	static readonly graphemeSegmenter: Intl.Segmenter | null =
-		typeof Intl !== "undefined" && Intl.Segmenter
-			? new Intl.Segmenter(undefined, {
-					granularity: "grapheme",
-				})
-			: null;
-
-	abstract getLine(): LyricLine;
-	abstract enable(time?: number, shouldPlay?: boolean): void;
-	abstract disable(): void;
-	abstract resume(): void;
-	abstract pause(): void;
-	onLineSizeChange(_size: [number, number]): void {}
-	setTransform(
-		top: number = this.top,
-		scale: number = this.scale,
-		opacity: number = this.opacity,
-		blur: number = this.blur,
-		_force = false,
-		delay = 0,
-		_mode: LyricLineRenderMode = LyricLineRenderMode.SOLID,
-	): void {
-		this.top = top;
-		this.scale = scale;
-		this.opacity = opacity;
-		this.blur = blur;
-		this.delay = delay;
-	}
-
-	rebuildElement(): void {}
-
-	/**
-	 * 判定歌词是否可以应用强调辉光效果
-	 *
-	 * 果子在对辉光效果的解释是一种强调（emphasized）效果
-	 *
-	 * 条件是一个单词时长大于等于 1s 且长度小于等于 7
-	 *
-	 * @param word 单词
-	 * @returns 是否可以应用强调辉光效果
-	 */
-	static shouldEmphasize(word: LyricWord): boolean {
-		if (isCJK(word.word)) return word.endTime - word.startTime >= 1000;
-
-		return (
-			word.endTime - word.startTime >= 1000 &&
-			word.word.trim().length <= 7 &&
-			word.word.trim().length > 1
-		);
-	}
-	abstract update(delta?: number): void;
-	dispose(): void {}
 }
